+++ README.md
@@ -15,8 +15,8 @@
     - [x] 우승자
     - [x] 참가자
   
-- [ ] 예외처리
-  - [ ] console로 처리되고 있던 예외처리 메세지를 api 응답으로 보내기
+- [x] 예외처리
+  - [x] console로 처리되고 있던 예외처리 메세지를 api 응답으로 보내기
   - [x] spring validaton을 이용해서 이름과 시도횟수의 blank 처리하기
