+++ src/main/java/racingcar/exception/invalidinput/InvalidInputException.java
@@ -1,0 +1,9 @@
+package racingcar.exception.invalidinput;
+
+public class InvalidInputException extends RuntimeException{
+
+    public InvalidInputException(String message) {
+        super(message);
+    }
+
+}
