+++ src/test/java/racingcar/domain/CarTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import racingcar.exception.CarNameLengthException;
+import racingcar.exception.invalidinput.CarNameLengthException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
